Refactor: Move search bar implementation to `exdock_search_bar.dart`
- Moved the search bar implementation from `overview_page_header.dart` to a new `ExdockSearchBar` widget in `exdock_search_bar.dart`.
- Implemented a callback mechanism for search suggestions and submission in the new widget.
- Updated `overview_page_header.dart` to use the new `ExdockSearchBar` widget.
