+++ lib/widgets/overview_page/overview_page_header.dart
@@ -1,4 +1,5 @@
 import 'package:exdock_backend_client/globals/globals.dart';
+import 'package:exdock_backend_client/widgets/input/exdock_search_bar.dart';
 import 'package:exdock_backend_client/widgets/overview_page/bulk/bulk_action.dart';
 import 'package:exdock_backend_client/widgets/overview_page/content/columns/overview_page_column.dart';
 import 'package:exdock_backend_client/widgets/overview_page/filters/filter.dart';
@@ -40,50 +41,15 @@ class _OverviewPageHeaderState extends State<OverviewPageHeader> {
           mainAxisSize: MainAxisSize.max,
           crossAxisAlignment: CrossAxisAlignment.center,
           children: [
-            SearchAnchor(
-              shrinkWrap: true,
-              viewBackgroundColor: Theme.of(context).indicatorColor,
-              builder: (context, searchController) {
-                return SearchBar(
-                  backgroundColor: WidgetStatePropertyAll(
-                    Theme.of(context).indicatorColor,
-                  ),
-                  controller: searchController,
-                  onTap: () {
-                    searchController.openView();
-                  },
-                  onChanged: (_) {
-                    searchController.openView();
-                  },
-                  onSubmitted: (searchText) {
-                    applyKeywordFilter(searchText);
-                  },
-                  padding: const WidgetStatePropertyAll(EdgeInsets.all(16)),
-                  trailing: const [Icon(Icons.search_rounded)],
-                );
-              },
-              suggestionsBuilder: (context, searchController) {
-                Function() onSelect(String searchText) {
-                  return () {
-                    applyKeywordFilter(searchText);
-                    searchController.closeView(searchText);
-                  };
-                }
-
-                ListTile createListTile(String searchText) {
-                  return ListTile(
-                    title: Text(searchText),
-                    onTap: onSelect(searchText),
-                  );
-                }
-
-                // TODO: get search suggestions
+            ExdockSearchBar(
+              onSubmitted: applyKeywordFilter,
+              getSearchSuggestions: (searchText) {
                 return [
-                  createListTile("search result 1"),
-                  createListTile("search result 2"),
-                  createListTile("search result 3"),
-                  createListTile("search result 4"),
-                  createListTile("search result 5"),
+                  "search result 1",
+                  "search result 2",
+                  "search result 3",
+                  "search result 4",
+                  "search result 5",
                 ];
               },
             ),
